+++ examples/smart_scraper_graph/ollama/smart_scraper_schema_ollama.py
@@ -1,0 +1,50 @@
+"""
+Basic example of scraping pipeline using SmartScraper with schema
+"""
+
+import json
+
+from pydantic import BaseModel, Field
+
+from scrapegraphai.graphs import SmartScraperGraph
+from scrapegraphai.utils import prettify_exec_info
+
+
+# ************************************************
+# Define the configuration for the graph
+# ************************************************
+class Project(BaseModel):
+    title: str = Field(description="The title of the project")
+    description: str = Field(description="The description of the project")
+
+
+class Projects(BaseModel):
+    projects: list[Project]
+
+
+graph_config = {
+    "llm": {"model": "ollama/llama3.2", "temperature": 0, "model_tokens": 4096},
+    "verbose": True,
+    "headless": False,
+}
+
+# ************************************************
+# Create the SmartScraperGraph instance and run it
+# ************************************************
+
+smart_scraper_graph = SmartScraperGraph(
+    prompt="List me all the projects with their description",
+    source="https://perinim.github.io/projects/",
+    schema=Projects,
+    config=graph_config,
+)
+
+result = smart_scraper_graph.run()
+print(json.dumps(result, indent=4))
+
+# ************************************************
+# Get graph execution info
+# ************************************************
+
+graph_exec_info = smart_scraper_graph.get_execution_info()
+print(prettify_exec_info(graph_exec_info))